AgriConnect-main/AgriConnect-main/backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-Cors==3.0.10
PyJWT==2.6.0
python-dotenv==1.0.0
google-generativeai==0.3.1
cryptography==40.0.2
passlib==1.7.4 